pyproject.toml
[tool.poetry]
name = "mirai-pixiv-bot"
version = "0.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
PixivPy = "^3.5.9"
json5 = "^0.9.5"
Pillow = "^7.2.0"
aiofiles = "^0.5.0"
loguru = "^0.5.1"
graia-application-mirai = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

